support bigtiff

+++ src/DEM.h
@@ -27,7 +27,7 @@ class DEM
 {
 protected:
 	float* pDem;
-	int width, height;
+	long long int width, height;
 public:
 	DEM()
 	{
